Diamond
Made it so that the crystals can now be collected, and they will be stored even after the application is closed. In general, I thought to use PlayerPrefs, but for this task it may work in the short term, but this is not the best approach for permanently storing data in the game. For a super casual game, it is possible to use Unity's built-in serialization system to store the number of crystals the player has collected. Use the BinaryFormatter class to serialize and deserialize data, and then save it to a binary file.