Tested and implemented ads and ad counter [prototype is complete]
Now the prototype of the game has ads, and it will not be shown to the player all the time, but at a certain frequency, and the IUnityAdsListener interface has also been added. At this point, the prototype is complete. Further, it is possible to implement where crystals can be spent (at this stage, the player simply collects them), or else come up with something else. I'll add a video to the description.